refactor: rename router file for consistency
