+++ frontend/src/main.tsx
@@ -2,7 +2,7 @@ import './index.css'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { RouterProvider } from "react-router/dom";
-import router from "./router/router.tsx";
+import router from "./router";
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
